refactor(web): organize components by feature domain
